[zlib] fixed warning C4244: conversion, possible loss of data

+++ Source/3rd Party/zlib/contrib/minizip/zip.c
@@ -1368,8 +1368,9 @@ local int zip64FlushWriteBuffer(zip64_internal* zi)
     {
 #ifndef NOCRYPT
         uInt i;
-        int t;
-        for (i=0;i<zi->ci.pos_in_buffered_data;i++)
+        Byte t;
+
+        for (i = 0; i < zi->ci.pos_in_buffered_data; i++)
             zi->ci.buffered_data[i] = zencode(zi->ci.keys, zi->ci.pcrc_32_tab, zi->ci.buffered_data[i],t);
 #endif
     }
